Refactor: status별 utils 분리
